backend/requirements.txt
# ============================================================
# HealthVault AI — Backend Dependencies
# ============================================================

# Core framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Supabase
supabase==2.4.3
python-jose[cryptography]==3.3.0

# Pydantic + settings
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Authentication
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# File handling + OCR
pillow==10.3.0
pytesseract==0.3.10
pypdf2==3.0.1
pymupdf==1.24.3            # PDF → image rendering (no poppler needed)
pdf2image==1.17.0          # fallback PDF renderer (requires poppler)
python-magic==0.4.27
aiofiles==23.2.1
boto3==1.34.84             # AWS S3 (fallback storage)

# AI / LLM
openai==1.25.0
google-generativeai==0.5.3
langchain==0.1.20
langchain-openai==0.1.6
tiktoken==0.6.0

# Scheduler
apscheduler==3.10.4

# WhatsApp / Notifications
twilio==9.0.5

# HTTP client
httpx==0.27.0
aiohttp==3.9.5

# Utilities
python-dotenv==1.0.1
orjson==3.10.3
structlog==24.1.0
tenacity==8.2.3            # retry logic

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
factory-boy==3.3.0
aiosqlite==0.20.0          # in-memory SQLite for tests
